Project/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
requests==2.26.0
python-dotenv==0.19.0
Flask-Mail==0.9.1
werkzeug==2.0.1
newspaper3k==0.2.8
python-dateutil==2.8.2 